+++ changelog.md
@@ -2,6 +2,7 @@ Upcoming
 ------
 * Added new manager methods for accessing groups, discussions, comments and users.
 * Added an `AppConfig` in apps.py for ease of configuration.
+* Added new manager methods to return active groups, discussions, and comments.
 
 v0.3.1
 ------